requirements.txt
asgiref==3.7.2
certifi==2024.6.2
charset-normalizer==3.3.2
colorama==0.4.6
distlib==0.3.8
Django==5.0.2
django-cors-headers==4.3.1
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.1
filelock==3.13.1
fsspec==2024.6.0
idna==3.7
intel-openmp==2021.4.0
Jinja2==3.1.4
llvmlite==0.42.0
MarkupSafe==2.1.5
mkl==2021.4.0
more-itertools==10.2.0
mpmath==1.3.0
networkx==3.3
numba==0.59.1
numpy==1.26.4
openai-whisper @ git+https://github.com/openai/whisper.git@ba3f3cd54b0e5b8ce1ab3de13e32122d0d5f98ab
platformdirs==4.1.0
PyJWT==2.8.0
python-dotenv==1.0.1
pytz==2024.1
regex==2024.5.15
requests==2.32.3
sqlparse==0.4.4
sympy==1.12.1
tbb==2021.12.0
tiktoken==0.7.0
torch==2.3.1
tqdm==4.66.4
typing_extensions==4.12.2
tzdata==2023.4
urllib3==2.2.1
virtualenv==20.25.0
streamlit
ffmpeg-python
sounddevice
pyttsx3
groq
scipy